Introduce view wrapping to allow for customization
Introduces a setting that points to a function that enables the
developer to wrap the views in this package.

The usecase in pursuit at the moment is to allow the integrating
django project to apply permissions to the views via the wrapper
function.

This works well enough to provide the needed functionality without
needing to repeat the entire set of urls.